pyproject.toml
[tool.poetry]
name = "event_map"
version = "0.1.0"
description = ""
authors = ["Vasilii Safronov <vasi1988@mail.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-Login = "^0.6.1"
Flask-Migrate = "^3.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-WTF = "^1.0.1"
email-validator = "^1.2.1"
psycopg2-binary = "^2.9.3"
pymongo = "^4.1.1"
requests = "^2.28.0"
gunicorn = "^20.1.0"
types-requests = "^2.27.30"
mypy = "^0.961"
gevent = "^21.12.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"
aiohttp = "^3.8.1"
backoff = "^2.1.2"
SQLAlchemy = "^1.4.39"
alembic = "^1.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
